+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v ?.?.? - ?? ??? ????
 - Removed [[nodiscard]] from ModelLoader::draw() to silence warnings.  Currently method always returns true so the return value isn't applicable.
+- Changed format of extensions listing header
 
 ## v 5.2.0 - 15 Aug 2024
 - OpenGLEngine tracks the state of the Caps Lock and Num Lock keys during key callbacks